Move RandomText to its own file
